Moving models to main project
